+++ pkg/webhook/manager.go
@@ -13,6 +13,7 @@ import (
 	"github.com/telekom/k8s-breakglass/pkg/breakglass/debug"
 	"github.com/telekom/k8s-breakglass/pkg/cert"
 	"github.com/telekom/k8s-breakglass/pkg/cli"
+	"github.com/telekom/k8s-breakglass/pkg/cluster"
 	"github.com/telekom/k8s-breakglass/pkg/indexer"
 	"github.com/telekom/k8s-breakglass/pkg/utils"
 	"go.uber.org/zap"
@@ -43,6 +44,7 @@ func Setup(
 	log *zap.SugaredLogger,
 	scheme *runtime.Scheme,
 	wc *cli.WebhookConfig,
+	ccProvider *cluster.ClientProvider,
 	enableValidatingWebhooks bool,
 	enableHTTP2 bool,
 	enableCertGeneration bool,
@@ -153,7 +155,7 @@ func Setup(
 				Client:       mgr.GetClient(),
 				Log:          log,
 				Decoder:      admission.NewDecoder(mgr.GetScheme()),
-				DebugHandler: debug.NewKubectlDebugHandler(mgr.GetClient(), nil),
+				DebugHandler: debug.NewKubectlDebugHandler(mgr.GetClient(), debug.AdaptClusterClientProvider(ccProvider)),
 			},
 		}